// frontend/src/lib/api/generated.ts
export interface components {
    schemas: {
        Actor: {
            readonly id: number;
            /** Format: email */
            readonly email: string;
            readonly full_name: string;
            readonly role: string;
        };
        Application: {
            readonly id: number;
            readonly owner: components["schemas"]["Actor"];
            title: string;
            category: components["schemas"]["ApplicationCategoryEnum"];
            description?: string;
            /** Format: decimal */
            amount?: string | null;
            readonly attachment_name: string | null;
            readonly attachment_download_url: string | null;
            readonly status: components["schemas"]["ApplicationStatusEnum"];
            readonly status_display: string;
            readonly transitions: components["schemas"]["StatusTransition"][];
            /** Format: date-time */
            readonly created_at: string;
            /** Format: date-time */
            readonly updated_at: string;
        };
        /**
         * @description * `NAME_CLEARANCE` - Name Clearance
         *     * `INCORPORATION` - Incorporation
         *     * `BO_DECLARATION` - Business Ownership Declaration
         *     * `DETAILS_UPDATE` - Details Update
         * @enum {string}
         */
        ApplicationCategoryEnum: "NAME_CLEARANCE" | "INCORPORATION" | "BO_DECLARATION" | "DETAILS_UPDATE";
// ...
        ApplicationStatusEnum: "DRAFT" | "SUBMITTED" | "UNDER_REVIEW" | "APPROVED" | "REJECTED";
// ...
        StatusTransition: {
            readonly id: number;
            readonly actor: components["schemas"]["Actor"];
            readonly from_status: components["schemas"]["ApplicationStatusEnum"];
            readonly to_status: components["schemas"]["ApplicationStatusEnum"];
            readonly comment: string;
            /** Format: date-time */
            readonly created_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}